+++ CHANGES.rst
@@ -5,6 +5,10 @@
   (`Issue #18 <https://github.com/flake8-commas/flake8-commas/issue/18>`_)
 - If there is an empty, tuple, list, dict, or function, do not report an error.
   (`Issue #17 <https://github.com/flake8-commas/flake8-commas/issue/17>`_)
+- Support PEP 3132 Python 3.5+ extended unpacking.
+  (`Issue #26 <https://github.com/flake8-commas/flake8-commas/issue/26>`_)
+- `*args` should not require a trailing comma.
+  (`Issue #27 <https://github.com/flake8-commas/flake8-commas/issue/27>`_)
 
 
 0.2.0 (2017-01-13)